API

Mods.is_invertible — Function

is_invertible(x::Mod) determines if x is invertible.

Mods.value — Function

value(a::Mod) returns the value of this Mod number.

julia> a = Mod{13}(11);

julia> value(a)
11

Mods.Mod — Type

Mod{m}(v) creates a modular number in mod m with value mod(v,m). Mod{m}() is equivalent to Mod{m}(0).

Mods.CRT — Function

CRT(m1,m2,...): Chinese Remainder Theorem

julia> CRT( Mod(4,11), Mod(8,14) )
Mods.Mod(92,154)

julia> 92%11
4

julia> 92%14
8

Mods.modulus — Function

modulus(a::Mod) returns the modulus of this Mod number.

julia> a = Mod{13}(11);

julia> modulus(a)
13

Mods.GaussMod — Type

GaussMod{m}(v) creates a modular number in mod m with value mod(v,m). GaussMod{m}() is equivalent to GaussMod{m}(0).

Base.inv — Function

inv(x::Mod) gives the multiplicative inverse of x. This may be abbreviated by x'.